+++ src/examples/plot_uncertainty.py
@@ -45,7 +45,7 @@ for sample in samples:
     sample = sample + values
     params = dict(zip(labels, sample))
     states.append(
-        neospy.CometElements(obj.desig, obj.covariance.epoch, **params).as_state
+        neospy.CometElements(obj.desig, obj.covariance.epoch, **params).state
     )
 
 # Propagate the position of all states to all time steps, recording the V mags