+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Restructured the Rust FOVs to be organized by observatory.
+- Renamed 'SpiceKernels.cache_kernel_reload' to 'kernel_reload' and changed it's args.
 
 ## [0.2.2] - 2024 - 5 - 20