+++ CHANGELOG.md
@@ -24,6 +24,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Restructured the Rust FOVs to be organized by observatory.
 - Renamed `SpiceKernels.cache_kernel_reload` to `kernel_reload` and changed it's args.
 
+### Fixed
+
+- Astropy will no longer warning about deprecated WCS header values for NEOWISE images.
+
 ## [0.2.2] - 2024 - 5 - 20
 
 ### Added